Renaming hearings reducers
