+++ client/app/hearings/index.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReduxBase from '@department-of-veterans-affairs/caseflow-frontend-toolkit/components/ReduxBase';
-import reducers, { initialState } from './reducers/reducers';
+import reducers from './reducers/index';
 import HearingsApp from './HearingsApp';
 
-const Hearings = (props) => <ReduxBase store={initialState} reducer={reducers}>
+const Hearings = (props) => <ReduxBase store={{}} reducer={reducers}>
   <HearingsApp {...props} />
 </ReduxBase>;
